Test and implement AnimationFactory
Test and implement AnimationFactory

+++ src/soliloquy/specs/graphics/assets/AnimationFrameSnippet.java
@@ -4,8 +4,8 @@ package soliloquy.specs.graphics.assets;
  * <b>AnimationFrameSnippet</b>
  * <p>
  * This class defines the state of an Animation at a given frame: It points to an {@link Image},
- * defines the boundaries for which to take a snippet from that Image, and defines the pixel
- * offsets for that snippet.
+ * defines the boundaries for which to take a snippet from that Image, and defines the offsets for
+ * that snippet in terms of the total percentage of the screen.
  * <p>
  * It is intended for this interface to be implemented anonymously at runtime.
  *
@@ -15,12 +15,12 @@ package soliloquy.specs.graphics.assets;
  */
 public interface AnimationFrameSnippet extends AssetSnippet {
     /**
-     * @return The x offset of the snippet in pixels
+     * @return The x offset of the snippet in the percentage of the screen
      */
-    int offsetX();
+    float offsetX();
 
     /**
-     * @return The y offset of the snippet in pixels
+     * @return The y offset of the snippet in the percentage of the screen
      */
-    int offsetY();
+    float offsetY();
 }
